[Subscrum]SCRUM-1: Add debug utils Remove chores Center queries generation for decoder head
