+++ p3former/segmentors/p3former.py
@@ -1,11 +1,4 @@
-'''testing only'''
-import matplotlib.pyplot as plt
-from datetime import datetime
-'''testing only'''
-
-import numpy as np
 import torch
-import torch.nn.functional as F
 from mmdet3d.registry import MODELS
 from mmdet3d.models.segmentors.cylinder3d import Cylinder3D
 from mmdet3d.structures import PointData
@@ -73,8 +66,6 @@ class _P3Former(Cylinder3D):
             offset_loss = self.offset_loss(pred_offsets, batch_data_samples)
             losses['offset_loss'] = sum(offset_loss)
 
-        # Validate offset
-        # validate_offset(pred_offsets, batch_inputs_dict, batch_data_samples, vis=True)
         for batch_i, p in enumerate(batch_inputs_dict['points']):
             batch_inputs_dict['points'][batch_i] = p[:,:3]
         assert len(pred_offsets[0]) == len(batch_inputs_dict['points'][0])
@@ -129,203 +120,3 @@ class _P3Former(Cylinder3D):
         for i in range(len(loss_list_list)):
             mean_loss_list.append(torch.mean(torch.stack(loss_list_list[i])))
         return mean_loss_list
-
-    def center_generation(self, batch_inputs_dict, batch_data_samples, embedding, is_train=True):
-        # Get heatmap
-        heatmap = []
-        x_edges_list = []
-        y_edges_list = []
-        for i, emb in enumerate(embedding):
-            if is_train:
-                valid = batch_data_samples[i].gt_pts_seg.pts_valid
-            else:
-                print("Not implemented vlaid for testing")
-                raise NotImplementedError
-            shifted_points = emb.detach()
-            shifted_points = shifted_points.cpu().numpy()[valid]
-
-            # Generate 2D heatmap
-            min_val = np.min(emb.detach().cpu().numpy(), axis=0)
-            max_val = np.max(emb.detach().cpu().numpy(), axis=0)
-            grid_size = 0.1
-            num_bins = np.ceil((max_val - min_val) / grid_size).astype(int)[:2]
-            H, x_edges, y_edges = np.histogram2d(shifted_points[:, 0], shifted_points[:, 1], bins=num_bins)
-
-            heatmap.append(H)
-            x_edges_list.append(x_edges)
-            y_edges_list.append(y_edges)
-
-        window_size = 5
-        threshold = 1
-        heatmap_pooled = F.max_pool2d(torch.tensor(heatmap), window_size, stride=5, padding=window_size//2)
-        heatmap_pooled = F.interpolate(heatmap_pooled.unsqueeze(0), 
-                                       size=(heatmap[0].shape[0], 
-                                             heatmap[0].shape[1]), mode='bilinear').squeeze(0)
-        
-        #  create binary mask
-        heatmap_pooled[heatmap_pooled >= threshold] = 1
-        heatmap_pooled[heatmap_pooled < threshold] = 0
-
-        #  Find contour on heatmap_pooled
-        heatmap_pooled = heatmap_pooled.cpu().numpy()
-
-        #  Find contour on heatmap_pooled
-        from skimage import measure
-        contours = measure.find_contours(heatmap_pooled[0], 0.5)
-        
-        # In each contour, get the center pixel, other pixels map to 0
-        new_heatmap = np.zeros_like(heatmap_pooled)
-        for contour in contours:
-            center = np.mean(contour, axis=0)
-            center = np.round(center).astype(int)
-            new_heatmap[0][center[0], center[1]] = 1
-  
-        pcls = []
-        for i, h in enumerate(new_heatmap):
-            x_centers = (x_edges_list[i][:-1] + x_edges_list[i][1:]) / 2
-            y_centers = (y_edges_list[i][:-1] + y_edges_list[i][1:]) / 2
-
-            # Generate a grid of indices
-            x_indices, y_indices = np.indices(h.shape)
-
-            # Repeat each index according to the corresponding histogram value
-            x_repeat = np.repeat(x_indices, h.astype(int).flatten())
-            y_repeat = np.repeat(y_indices, h.astype(int).flatten())
-            z_val = np.mean(h)
-            z_repeat = np.full_like(x_repeat, z_val, dtype=float)
-
-            # Map the indices to the bin centers to get the point coordinates
-            pcl = np.column_stack((x_centers[x_repeat], y_centers[y_repeat], z_repeat))
-
-            pcls.append(pcl)
-        return torch.tensor(pcls).cuda()
-
-
-
-
-
-
-
-
-
-
-
-def validate_offset(pred_offsets, batch_inputs_dict, batch_data_samples, vis=True):
-    valid = batch_data_samples[0].gt_pts_seg.pts_valid
-    time = datetime.now().strftime("%Y%m%d%H%M%S")
-    pts = batch_inputs_dict['points']
-    draw_point(pts[0].cpu().numpy(), name=f'pcl_gt_img_{time}.png')
-    draw_point(pts[0].cpu().numpy()[valid], name=f'pcl_thing_img_{time}.png')
-
-    pts[0] = pts[0][:,:3]
-
-    embedding = [offset + xyz for offset, xyz in zip(pred_offsets, pts)]
-    for emb in embedding:
-        with torch.no_grad():
-            shifted_points = emb.cpu().numpy()
-            draw_point(shifted_points, name=f'shifted_pointcloud_{time}.png')
-            draw_point(shifted_points[valid], name=f'shifted_thing_pointcloud_{time}.png')
-
-
-def draw_point(pcl, indices=None, name='pcl_img.png', s=1):
-    # Assuming `embedding` is your point cloud and `indices` are the indices of the centroids
-    embedding_np = pcl  # Convert to numpy array for plotting
-    
-    # Create a new color array
-    colors = np.full(embedding_np.shape[0], 'w')  # All points are blue
-    
-    # Create a new size array
-    sizes = np.full(embedding_np.shape[0], s)  # All points are size 1
-    
-    if indices is not None:
-        indices_np = indices.flatten()  # Flatten the indices array for indexing
-        colors[indices_np] = 'r'  # Centroid points are red
-        sizes[indices_np] = 10  # Centroid points are size 20
-
-    # Create a 3D plot
-    fig = plt.figure(figsize=(30, 30))
-    ax = fig.add_subplot()
-
-    # Plot the point cloud in 3D
-    scatter = ax.scatter(embedding_np[:, 0], embedding_np[:, 1], c=colors, s=sizes)
-
-    # Set background color to black
-    ax.set_facecolor('black')
-    ax.set_aspect('auto')
-    # ax.set_axis_off()
-
-    # Hide grid
-    # ax.grid(False)
-
-    # Save the plot to an image file
-    plt.savefig(name)
-
-def write_img(img, name='heatmap.png'):
-    plt.imsave(name, img)
-
-def draw_point_with(pcl, indices=None, name='pcl_img.png', s=1, pcl2=None, s2=1):
-    # Assuming `embedding` is your point cloud and `indices` are the indices of the centroids
-    embedding_np = pcl  # Convert to numpy array for plotting
-    
-    # Create a new color array
-    colors = np.full(embedding_np.shape[0], 'w')  # All points are blue
-    
-    # Create a new size array
-    sizes = np.full(embedding_np.shape[0], s)  # All points are size 1
-    
-    if indices is not None:
-        indices_np = indices.flatten()  # Flatten the indices array for indexing
-        colors[indices_np] = 'b'  # Centroid points are red
-        sizes[indices_np] = 10  # Centroid points are size 20
-
-    # Create a 3D plot
-    fig = plt.figure(figsize=(30, 30))
-    ax = fig.add_subplot()
-
-    
-
-    # Plot the point cloud in 3D
-    scatter = ax.scatter(embedding_np[:, 0], embedding_np[:, 1], c=colors, s=sizes)
-    if pcl2 is not None:
-        scatter = ax.scatter(pcl2[:, 0], pcl2[:, 1], c='r', s=s2)
-
-    # Set background color to black
-    ax.set_facecolor('black')
-    ax.set_aspect('auto')
-    # ax.set_axis_off()
-
-    # Hide grid
-    # ax.grid(False)
-
-    # Save the plot to an image file
-    plt.savefig(name)
-
-
-def auto_quantize_point_cloud(points, num_levels):
-    """
-    Automatically quantize a point cloud based on the number of quantization levels.
-    
-    Parameters:
-    points (np.ndarray): The point cloud data as an (N, 3) array.
-    num_levels (int): The number of quantization levels per dimension.
-    
-    Returns:
-    np.ndarray: The quantized point cloud as an (N, 3) array.
-    """
-    # Determine the bounding box of the point cloud
-    min_coords = np.min(points, axis=0)
-    max_coords = np.max(points, axis=0)
-    
-    # Compute the range of the point cloud in each dimension
-    range_coords = max_coords - min_coords
-    
-    # Compute the grid size
-    grid_size = range_coords / num_levels
-    
-    # Quantize the points
-    quantized_points = np.floor((points - min_coords) / grid_size + 0.5).astype(np.int32)
-    
-    # Reconstruct the quantized points
-    quantized_points = quantized_points * grid_size + min_coords
-    
-    return quantized_points
